refactor: add sql parser type druid as constant (#2156)
