+++ sqlparser/seata-sqlparser-core/src/main/java/io/seata/sqlparser/SqlParserType.java
@@ -13,20 +13,17 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package io.seata.sqlparser.druid;
-
-import com.alibaba.druid.util.JdbcUtils;
-import io.seata.common.loader.LoadLevel;
-import io.seata.sqlparser.SqlParserType;
-import io.seata.sqlparser.util.DbTypeParser;
+package io.seata.sqlparser;
 
 /**
  * @author ggndnn
  */
-@LoadLevel(name = SqlParserType.SQL_PARSER_TYPE_DRUID)
-public class DruidDbTypeParser implements DbTypeParser {
-    @Override
-    public String parseFromJdbcUrl(String jdbcUrl) {
-        return JdbcUtils.getDbType(jdbcUrl, null);
+public final class SqlParserType {
+    private SqlParserType() {
     }
+
+    /**
+     * The constant SQL_PARSER_TYPE_DRUID.
+     */
+    public static final String SQL_PARSER_TYPE_DRUID = "druid";
 }
